+++ engine/wolf_engine_core/src/engine_builder.rs
@@ -46,3 +46,27 @@ impl EngineBuilder {
         Ok((event_loop, context))
     }
 }
+
+#[cfg(test)]
+mod engin_builder_tests {
+    use crate::events::MainEventSender;
+
+    #[test]
+    fn should_add_resources() {
+        let (_event_loop, context) = crate::init().with_resource(0).build().unwrap();
+
+        assert!(
+            context.resources().get::<i32>().is_ok(),
+            "The resources were not used"
+        );
+    }
+
+    #[test]
+    fn should_add_event_sender_resource_by_default() {
+        let (_event_loop, context) = crate::init().build().unwrap();
+        let _event_sender = context
+            .resources()
+            .get_mut::<MainEventSender>()
+            .expect("No event sender was added.");
+    }
+}
